test: migrate example app e2e tests to Storybook
