+++ package.json
@@ -16,9 +16,9 @@
     "postbuild": "yarn assets",
     "build:app": "nx build example-app",
     "build:storybook": "nx run ngx-foundation-sites:build-storybook",
-    "e2e": "nx run example-app-e2e:e2e",
-    "e2e:ci": "nx run example-app-e2e:e2e:ci-development",
-    "e2e:watch": "nx run example-app-e2e:watch",
+    "e2e": "nx run ngx-foundation-sites-e2e:e2e",
+    "e2e:ci": "nx run ngx-foundation-sites-e2e:e2e:ci",
+    "e2e:watch": "nx run ngx-foundation-sites-e2e:e2e:watch",
     "format": "nx format:write",
     "lint": "nx run-many --target=lint --all",
     "start": "nx serve example-app",
