--- packages/example-app/example-app-e2e/.eslintrc.json
@@ -1,10 +1,0 @@
-{
-  "extends": ["plugin:cypress/recommended", "../../../.eslintrc.json"],
-  "ignorePatterns": ["!**/*"],
-  "overrides": [
-    {
-      "files": ["*.ts", "*.tsx", "*.js", "*.jsx"],
-      "rules": {}
-    }
-  ]
-}
