--- packages/example-app/example-app-e2e/README.md
@@ -1,5 +1,0 @@
-| Description                              | Angular development mode                    | Angular production mode                    |
-| ---------------------------------------- | ------------------------------------------- | ------------------------------------------ |
-| Open Cypress app in watch mode           | `nx run example-app-e2e:watch`              | `nx run example-app-e2e:watch:production`  |
-| Run end-to-end tests once in Cypress app | `nx run example-app-e2e:e2e`                | `nx run example-app-e2e:e2e:production`    |
-| Run end-to-end tests headless once       | `nx run example-app-e2e:e2e:ci-development` | `nx run example-app-e2e:e2e:ci-production` |
