--- packages/example-app/example-app-e2e/project.json
@@ -1,67 +1,0 @@
-{
-  "name": "example-app-e2e",
-  "$schema": "../../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "packages/example-app/example-app-e2e/src",
-  "projectType": "application",
-  "targets": {
-    "e2e": {
-      "executor": "@nrwl/cypress:cypress",
-      "options": {
-        "cypressConfig": "packages/example-app/example-app-e2e/cypress.config.ts",
-        "headed": true,
-        "testingType": "e2e"
-      },
-      "configurations": {
-        "ci-development": {
-          "cypressConfig": "packages/example-app/example-app-e2e/cypress.ci.config.ts",
-          "devServerTarget": "example-app:serve:e2e-development",
-          "headed": false
-        },
-        "ci-production": {
-          "cypressConfig": "packages/example-app/example-app-e2e/cypress.ci.config.ts",
-          "devServerTarget": "example-app:serve:e2e-production",
-          "headed": false
-        },
-        "development": {
-          "devServerTarget": "example-app:serve:e2e-development"
-        },
-        "production": {
-          "devServerTarget": "example-app:serve:e2e-production"
-        },
-        "watch": {
-          "watch": true
-        }
-      },
-      "defaultConfiguration": "development"
-    },
-    "lint": {
-      "executor": "@nrwl/linter:eslint",
-      "outputs": ["{options.outputFile}"],
-      "options": {
-        "lintFilePatterns": [
-          "packages/example-app/example-app-e2e/**/*.{js,ts}"
-        ]
-      }
-    },
-    "watch": {
-      "executor": "@nrwl/cypress:cypress",
-      "options": {
-        "cypressConfig": "packages/example-app/example-app-e2e/cypress.config.ts",
-        "headed": true,
-        "watch": true,
-        "testingType": "e2e"
-      },
-      "configurations": {
-        "development": {
-          "devServerTarget": "example-app:serve:e2e-development"
-        },
-        "production": {
-          "devServerTarget": "example-app:serve:e2e-production"
-        }
-      },
-      "defaultConfiguration": "development"
-    }
-  },
-  "tags": [],
-  "implicitDependencies": ["example-app"]
-}
