--- packages/example-app/example-app-e2e/src/e2e/app.cy.ts
@@ -1,9 +1,0 @@
-import * as App from '../support/app.po';
-
-describe('Example app', () => {
-  beforeEach(() => App.navigateTo());
-
-  it('displays welcome message', () => {
-    App.getTitle().should('contain.text', 'Foundation for Angular Sites');
-  });
-});
