--- packages/example-app/example-app-e2e/src/support/app.po.ts
@@ -1,2 +1,0 @@
-export const getTitle = () => cy.get('example-app > h1');
-export const navigateTo = () => cy.visit('/');
