--- packages/example-app/example-app-e2e/src/support/tabs.po.ts
@@ -1,7 +1,0 @@
-import * as loader from '@jscutlery/cypress-harness';
-import { FasTabHarness } from 'ngx-foundation-sites/testing';
-
-export const getAllTabs = () => loader.getAllHarnesses(FasTabHarness);
-export const getTabWithTitle = (title: string) =>
-  loader.getHarness(FasTabHarness.with({ title }));
-export const navigateTo = () => cy.visit('/');
