--- packages/example-app/example-app-e2e/src/support/tabs/tabs.po.ts
@@ -1,1 +1,0 @@
-export const navigateTo = () => cy.visit('/containers/tabs');
