--- packages/example-app/example-app-e2e/tsconfig.json
@@ -1,26 +1,0 @@
-{
-  "extends": "../../../tsconfig.base.json",
-  "compilerOptions": {
-    "sourceMap": false,
-    "outDir": "../../../dist/out-tsc",
-    "allowJs": true,
-    "types": ["@testing-library/cypress", "cypress", "node"],
-    "forceConsistentCasingInFileNames": true,
-    "strict": true,
-    "noImplicitOverride": true,
-    "noPropertyAccessFromIndexSignature": true,
-    "noImplicitReturns": true,
-    "noFallthroughCasesInSwitch": true
-  },
-  "include": [
-    "src/**/*.ts",
-    "src/**/*.js",
-    "cypress.config.ts",
-    "cypress.ci.config.ts"
-  ],
-  "angularCompilerOptions": {
-    "strictInjectionParameters": true,
-    "strictInputAccessModifiers": true,
-    "strictTemplates": true
-  }
-}
