+++ packages/ngx-foundation-sites-e2e/README.md
@@ -1,0 +1,5 @@
+| Description                              | Command                                     |
+| ---------------------------------------- | ------------------------------------------- |
+| Open Cypress app in watch mode           | `nx run ngx-foundation-sites-e2e:e2e:watch` |
+| Run end-to-end tests once in Cypress app | `nx run ngx-foundation-sites-e2e:e2e`       |
+| Run end-to-end tests headless once       | `nx run ngx-foundation-sites-e2e:e2e:ci`    |
