+++ packages/ngx-foundation-sites-e2e/cypress.ci.config.ts
@@ -1,13 +1,15 @@
 import { defineConfig } from 'cypress';
-import { nxE2EPreset } from '@nrwl/cypress/plugins/cypress-preset';
+import { nxE2EStorybookPreset } from '@nrwl/storybook/presets/cypress';
 import { getPreprocessorConfig } from '@jscutlery/cypress-harness/preprocessor-config';
 
 export default defineConfig({
   e2e: {
-    ...nxE2EPreset(__dirname),
-    video: true,
+    ...nxE2EStorybookPreset(__dirname),
+    screenshotOnRunFailure: false,
+    video: false,
     viewportHeight: 1080,
     viewportWidth: 1920,
+    watchForFileChanges: false,
     ...getPreprocessorConfig(),
   },
 });
